Human diffusion processing won't try to copy labels if they don't exist.

+++ human_diffusion_preprocessing_MRtrix.bash
@@ -574,7 +574,9 @@ labels=${work_dir}/${id}_IITmean_RPI_labels.nii.gz;
 
 if [[ ! -e ${labels} ]];then
 	source_labels=${BIGGUS_DISKUS}/../mouse/VBM_25ADNI01_IITmean_RPI-results/connectomics/${id}/${id}_IITmean_RPI_labels.nii.gz;
-	cp ${source_labels} ${labels};
+	if [[ -e ${source_labels} ]];then
+		cp ${source_labels} ${labels};
+	fi
 	if [[ ! -e ${labels} ]];then
 		echo "Process stopped at ${stage}.";
 		echo "SAMBA labels do not exist yet."
